Move the template filters into the ui blueprint
